pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "brag"
version = "0.1.0"
description = "A CLI tool to track your work accomplishments and build your brag document"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Theau Poulat", email = "theau.poulat@gmail.com" }
]
keywords = ["brag", "accomplishments", "cli", "productivity"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Utilities",
]
dependencies = [
    "click>=8.0.0",
    "questionary>=2.0.0",
    "pyyaml>=6.0.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "freezegun>=1.4.0",
]

[project.scripts]
brag = "brag.cli:main"

[project.urls]
Homepage = "https://github.com/theaupoulat/brag"
Repository = "https://github.com/theaupoulat/brag"

[tool.hatch.build.targets.wheel]
packages = ["src/brag"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/README.md",
    "/LICENSE",
]
